+++ UIComponents.Generators/Generators/Property/Inputs/UICGeneratorInputMultiline.cs
@@ -26,9 +26,6 @@ public class UICGeneratorInputMultiline : UICGeneratorProperty
         };
         input.Value = args.PropertyValue?.ToString()?? string.Empty;
 
-        if (args.Configuration.TryGetPermissionService(out var permissionService))
-            input.Readonly = !(await permissionService!.CanEditObject(args.ClassObject!) && await permissionService.CanEditProperty(args.ClassObject, args.PropertyName));
-
 
         if (args.Options.CheckClientSideValidation)
         {